Set up abstraction for both interactions and text commands
